feat(primitives): expose advanced operator sinks
- Move branch-added operator signals and sinks out of private nested types.

- Add public API baselines for primitive and reactive Advanced namespaces.
