+++ src/Primitives.Shared/Advanced/SynchronizeObjectSignal{T}.cs
@@ -1,0 +1,39 @@
+// Copyright (c) 2019-2026 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for full license information.
+
+#if REACTIVE_SHIM
+namespace ReactiveUI.Primitives.Reactive.Advanced;
+#else
+namespace ReactiveUI.Primitives.Advanced;
+#endif
+
+/// <summary>Object-gated <c>Synchronize</c> compatibility signal.</summary>
+/// <typeparam name="T">The value type.</typeparam>
+public sealed class SynchronizeObjectSignal<T> : IObservable<T>
+{
+    /// <summary>The source observable.</summary>
+    private readonly IObservable<T> _source;
+
+    /// <summary>The shared gate.</summary>
+    private readonly object _gate;
+
+    /// <summary>Initializes a new instance of the <see cref="SynchronizeObjectSignal{T}"/> class.</summary>
+    /// <param name="source">The source observable.</param>
+    /// <param name="gate">The gate shared across subscriptions and other synchronized sequences.</param>
+    public SynchronizeObjectSignal(IObservable<T> source, object gate)
+    {
+        _source = source;
+        _gate = gate;
+    }
+
+    /// <inheritdoc/>
+    public IDisposable Subscribe(IObserver<T> observer)
+    {
+        ArgumentExceptionHelper.ThrowIfNull(observer);
+
+        SynchronizeObjectWitness<T> sink = new(observer, _gate);
+        sink.SetSubscription(_source.Subscribe(sink));
+        return sink;
+    }
+}
