+++ src/Primitives.Shared/SignalOperatorMixins.PassThroughSignals.cs
@@ -186,90 +186,4 @@ public static partial class LinqExtensions
             return sink;
         }
     }
-
-    /// <summary>Dedicated signal for object-gated <c>Synchronize</c> compatibility.</summary>
-    /// <typeparam name="T">The value type.</typeparam>
-    private sealed class SynchronizeObjectSignal<T> : IObservable<T>
-    {
-        /// <summary>The source observable.</summary>
-        private readonly IObservable<T> _source;
-
-        /// <summary>The shared gate.</summary>
-        private readonly object _gate;
-
-        /// <summary>Initializes a new instance of the <see cref="SynchronizeObjectSignal{T}"/> class.</summary>
-        /// <param name="source">The source observable.</param>
-        /// <param name="gate">The gate shared across subscriptions and other synchronized sequences.</param>
-        internal SynchronizeObjectSignal(IObservable<T> source, object gate)
-        {
-            _source = source;
-            _gate = gate;
-        }
-
-        /// <inheritdoc/>
-        public IDisposable Subscribe(IObserver<T> observer)
-        {
-            ArgumentExceptionHelper.ThrowIfNull(observer);
-
-            SynchronizeObjectWitness sink = new(observer, _gate);
-            sink.SetSubscription(_source.Subscribe(sink));
-            return sink;
-        }
-
-        /// <summary>Observer that serializes notifications using an object gate.</summary>
-        private sealed class SynchronizeObjectWitness : IObserver<T>, IDisposable
-        {
-            /// <summary>The downstream observer.</summary>
-            private readonly IObserver<T> _observer;
-
-            /// <summary>The gate that serializes every forwarded notification.</summary>
-            private readonly object _gate;
-
-            /// <summary>The upstream subscription.</summary>
-            private IDisposable? _subscription;
-
-            /// <summary>Initializes a new instance of the <see cref="SynchronizeObjectWitness"/> class.</summary>
-            /// <param name="observer">The downstream observer.</param>
-            /// <param name="gate">The gate shared with other synchronized observers.</param>
-            internal SynchronizeObjectWitness(IObserver<T> observer, object gate)
-            {
-                _observer = observer;
-                _gate = gate;
-            }
-
-            /// <inheritdoc/>
-            public void OnNext(T value)
-            {
-                lock (_gate)
-                {
-                    _observer.OnNext(value);
-                }
-            }
-
-            /// <inheritdoc/>
-            public void OnError(Exception error)
-            {
-                lock (_gate)
-                {
-                    _observer.OnError(error);
-                }
-            }
-
-            /// <inheritdoc/>
-            public void OnCompleted()
-            {
-                lock (_gate)
-                {
-                    _observer.OnCompleted();
-                }
-            }
-
-            /// <summary>Assigns the upstream subscription.</summary>
-            /// <param name="subscription">The upstream subscription.</param>
-            public void SetSubscription(IDisposable subscription) => SinkSubscription.Set(ref _subscription, subscription);
-
-            /// <inheritdoc/>
-            public void Dispose() => SinkSubscription.Dispose(ref _subscription);
-        }
-    }
 }
