+++ src/Primitives.Shared/SignalOperatorParityMixins.RxNames.cs
@@ -807,251 +807,4 @@ public static partial class LinqExtensions
             return new TaskChainSignal<T>(sources);
         }
     }
-
-    /// <summary>Observer that turns downstream <c>OnNext</c> exceptions into a terminal error and upstream disposal.</summary>
-    /// <typeparam name="T">The value type.</typeparam>
-    private sealed class SubscribeSafeObserver<T> : IObserver<T>, IDisposable
-    {
-        /// <summary>The wrapped observer.</summary>
-        private readonly IObserver<T> _observer;
-
-        /// <summary>The upstream subscription.</summary>
-        private readonly SingleReplaceableDisposable _subscription = new();
-
-        /// <summary>Non-zero after terminal notification or disposal.</summary>
-        private int _stopped;
-
-        /// <summary>Initializes a new instance of the <see cref="SubscribeSafeObserver{T}"/> class.</summary>
-        /// <param name="observer">The wrapped observer.</param>
-        internal SubscribeSafeObserver(IObserver<T> observer) => _observer = observer;
-
-        /// <inheritdoc/>
-        public void Dispose()
-        {
-            Interlocked.Exchange(ref _stopped, 1);
-            _subscription.Dispose();
-        }
-
-        /// <inheritdoc/>
-        public void OnCompleted()
-        {
-            if (Interlocked.Exchange(ref _stopped, 1) != 0)
-            {
-                return;
-            }
-
-            try
-            {
-                _observer.OnCompleted();
-            }
-            finally
-            {
-                _subscription.Dispose();
-            }
-        }
-
-        /// <inheritdoc/>
-        public void OnError(Exception error) => StopWithError(error);
-
-        /// <inheritdoc/>
-        public void OnNext(T value)
-        {
-            if (Volatile.Read(ref _stopped) != 0)
-            {
-                return;
-            }
-
-            try
-            {
-                _observer.OnNext(value);
-            }
-            catch (Exception error) when (!FatalExceptionHelper.IsFatal(error))
-            {
-                StopWithError(error);
-            }
-        }
-
-        /// <summary>Assigns the upstream subscription.</summary>
-        /// <param name="subscription">The upstream subscription.</param>
-        internal void SetSubscription(IDisposable subscription)
-        {
-            _subscription.Create(subscription);
-            if (Volatile.Read(ref _stopped) == 0)
-            {
-                return;
-            }
-
-            _subscription.Dispose();
-        }
-
-        /// <summary>Forwards an error exactly once and disposes the upstream subscription.</summary>
-        /// <param name="error">The terminal error.</param>
-        private void StopWithError(Exception error)
-        {
-            if (Interlocked.Exchange(ref _stopped, 1) != 0)
-            {
-                return;
-            }
-
-            try
-            {
-                _observer.OnError(error);
-            }
-            finally
-            {
-                _subscription.Dispose();
-            }
-        }
-    }
-
-    /// <summary>Signal that projects each value to an enumerable sequence and emits the enumerable values.</summary>
-    /// <typeparam name="TSource">The source value type.</typeparam>
-    /// <typeparam name="TResult">The result value type.</typeparam>
-    private sealed class SelectManyEnumerableSignal<TSource, TResult> : IObservable<TResult>
-    {
-        /// <summary>The source observable.</summary>
-        private readonly IObservable<TSource> _source;
-
-        /// <summary>The enumerable projection.</summary>
-        private readonly Func<TSource, IEnumerable<TResult>> _selector;
-
-        /// <summary>Initializes a new instance of the <see cref="SelectManyEnumerableSignal{TSource, TResult}"/> class.</summary>
-        /// <param name="source">The source observable.</param>
-        /// <param name="selector">The enumerable projection.</param>
-        internal SelectManyEnumerableSignal(IObservable<TSource> source, Func<TSource, IEnumerable<TResult>> selector)
-        {
-            _source = source;
-            _selector = selector;
-        }
-
-        /// <inheritdoc/>
-        public IDisposable Subscribe(IObserver<TResult> observer)
-        {
-            ArgumentExceptionHelper.ThrowIfNull(observer);
-
-            SelectManyEnumerableObserver<TSource, TResult> sink = new(observer, _selector);
-            sink.SetSubscription(_source.Subscribe(sink));
-            return sink;
-        }
-    }
-
-    /// <summary>Observer for enumerable <c>SelectMany</c>.</summary>
-    /// <typeparam name="TSource">The source value type.</typeparam>
-    /// <typeparam name="TResult">The result value type.</typeparam>
-    private sealed class SelectManyEnumerableObserver<TSource, TResult> : IObserver<TSource>, IDisposable
-    {
-        /// <summary>The downstream observer.</summary>
-        private readonly IObserver<TResult> _observer;
-
-        /// <summary>The enumerable projection.</summary>
-        private readonly Func<TSource, IEnumerable<TResult>> _selector;
-
-        /// <summary>The upstream subscription.</summary>
-        private readonly SingleReplaceableDisposable _subscription = new();
-
-        /// <summary>Non-zero after terminal notification or disposal.</summary>
-        private int _stopped;
-
-        /// <summary>Initializes a new instance of the <see cref="SelectManyEnumerableObserver{TSource, TResult}"/> class.</summary>
-        /// <param name="observer">The downstream observer.</param>
-        /// <param name="selector">The enumerable projection.</param>
-        internal SelectManyEnumerableObserver(IObserver<TResult> observer, Func<TSource, IEnumerable<TResult>> selector)
-        {
-            _observer = observer;
-            _selector = selector;
-        }
-
-        /// <inheritdoc/>
-        public void Dispose()
-        {
-            Interlocked.Exchange(ref _stopped, 1);
-            _subscription.Dispose();
-        }
-
-        /// <inheritdoc/>
-        public void OnCompleted()
-        {
-            if (Interlocked.Exchange(ref _stopped, 1) != 0)
-            {
-                return;
-            }
-
-            try
-            {
-                _observer.OnCompleted();
-            }
-            finally
-            {
-                _subscription.Dispose();
-            }
-        }
-
-        /// <inheritdoc/>
-        public void OnError(Exception error) => StopWithError(error);
-
-        /// <inheritdoc/>
-        public void OnNext(TSource value)
-        {
-            if (Volatile.Read(ref _stopped) != 0)
-            {
-                return;
-            }
-
-            try
-            {
-                var values = _selector(value);
-                if (values is null)
-                {
-                    StopWithError(new InvalidOperationException("SelectMany selector returned null."));
-                    return;
-                }
-
-                foreach (var result in values)
-                {
-                    if (Volatile.Read(ref _stopped) != 0)
-                    {
-                        return;
-                    }
-
-                    _observer.OnNext(result);
-                }
-            }
-            catch (Exception error) when (!FatalExceptionHelper.IsFatal(error))
-            {
-                StopWithError(error);
-            }
-        }
-
-        /// <summary>Assigns the upstream subscription.</summary>
-        /// <param name="subscription">The upstream subscription.</param>
-        internal void SetSubscription(IDisposable subscription)
-        {
-            _subscription.Create(subscription);
-            if (Volatile.Read(ref _stopped) == 0)
-            {
-                return;
-            }
-
-            _subscription.Dispose();
-        }
-
-        /// <summary>Forwards an error exactly once and disposes the upstream subscription.</summary>
-        /// <param name="error">The terminal error.</param>
-        private void StopWithError(Exception error)
-        {
-            if (Interlocked.Exchange(ref _stopped, 1) != 0)
-            {
-                return;
-            }
-
-            try
-            {
-                _observer.OnError(error);
-            }
-            finally
-            {
-                _subscription.Dispose();
-            }
-        }
-    }
 }
