+++ src/ReactiveUI.Primitives.Reactive/PublicAPI/net10.0/PublicAPI.Unshipped.txt
@@ -1,1 +1,52 @@
 #nullable enable
+
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedSignal<TSource, TResult>
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedSignal<TSource, TResult>.IsRequiredSubscribeOnCurrentThread() -> bool
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedSignal<TSource, TResult>.MapIndexedSignal(System.IObservable<TSource>! source, System.Func<TSource, int, TResult>! selector) -> void
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedSignal<TSource, TResult>.Subscribe(System.IObserver<TResult>! observer) -> System.IDisposable!
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedWitness<TSource, TResult>
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedWitness<TSource, TResult>.MapIndexedWitness(System.IObserver<TResult>! observer, System.Func<TSource, int, TResult>! selector) -> void
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedWitness<TSource, TResult>.OnCompleted() -> void
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedWitness<TSource, TResult>.OnError(System.Exception! error) -> void
+ReactiveUI.Primitives.Reactive.Advanced.MapIndexedWitness<TSource, TResult>.OnNext(TSource value) -> void
+ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentBlendCoordinator<T>
+ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentBlendCoordinator<T>.Dispose() -> void
+ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentBlendCoordinator<T>.MaxConcurrentBlendCoordinator(System.IObserver<T>! observer) -> void
+ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentBlendCoordinator<T>.Run(System.Collections.Generic.IEnumerable<System.IObservable<T>!>! sources, int maxConcurrent) -> ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentBlendCoordinator<T>!
+ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentEnumerableBlendSignal<T>
+ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentEnumerableBlendSignal<T>.MaxConcurrentEnumerableBlendSignal(System.Collections.Generic.IEnumerable<System.IObservable<T>!>! sources, int maxConcurrent) -> void
+ReactiveUI.Primitives.Reactive.Advanced.MaxConcurrentEnumerableBlendSignal<T>.Subscribe(System.IObserver<T>! observer) -> System.IDisposable!
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableObserver<TSource, TResult>
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableObserver<TSource, TResult>.Dispose() -> void
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableObserver<TSource, TResult>.OnCompleted() -> void
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableObserver<TSource, TResult>.OnError(System.Exception! error) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableObserver<TSource, TResult>.OnNext(TSource value) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableObserver<TSource, TResult>.SelectManyEnumerableObserver(System.IObserver<TResult>! observer, System.Func<TSource, System.Collections.Generic.IEnumerable<TResult>!>! selector) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableObserver<TSource, TResult>.SetSubscription(System.IDisposable! subscription) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableSignal<TSource, TResult>
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableSignal<TSource, TResult>.SelectManyEnumerableSignal(System.IObservable<TSource>! source, System.Func<TSource, System.Collections.Generic.IEnumerable<TResult>!>! selector) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SelectManyEnumerableSignal<TSource, TResult>.Subscribe(System.IObserver<TResult>! observer) -> System.IDisposable!
+ReactiveUI.Primitives.Reactive.Advanced.SubscribeSafeObserver<T>
+ReactiveUI.Primitives.Reactive.Advanced.SubscribeSafeObserver<T>.Dispose() -> void
+ReactiveUI.Primitives.Reactive.Advanced.SubscribeSafeObserver<T>.OnCompleted() -> void
+ReactiveUI.Primitives.Reactive.Advanced.SubscribeSafeObserver<T>.OnError(System.Exception! error) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SubscribeSafeObserver<T>.OnNext(T value) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SubscribeSafeObserver<T>.SetSubscription(System.IDisposable! subscription) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SubscribeSafeObserver<T>.SubscribeSafeObserver(System.IObserver<T>! observer) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectSignal<T>
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectSignal<T>.Subscribe(System.IObserver<T>! observer) -> System.IDisposable!
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectSignal<T>.SynchronizeObjectSignal(System.IObservable<T>! source, object! gate) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectWitness<T>
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectWitness<T>.Dispose() -> void
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectWitness<T>.OnCompleted() -> void
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectWitness<T>.OnError(System.Exception! error) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectWitness<T>.OnNext(T value) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectWitness<T>.SetSubscription(System.IDisposable! subscription) -> void
+ReactiveUI.Primitives.Reactive.Advanced.SynchronizeObjectWitness<T>.SynchronizeObjectWitness(System.IObserver<T>! observer, object! gate) -> void
+ReactiveUI.Primitives.Reactive.Advanced.TaskChainCoordinator<T>
+ReactiveUI.Primitives.Reactive.Advanced.TaskChainCoordinator<T>.Dispose() -> void
+ReactiveUI.Primitives.Reactive.Advanced.TaskChainCoordinator<T>.Run(System.IObservable<System.Threading.Tasks.Task<T>!>! sources) -> ReactiveUI.Primitives.Reactive.Advanced.TaskChainCoordinator<T>!
+ReactiveUI.Primitives.Reactive.Advanced.TaskChainCoordinator<T>.TaskChainCoordinator(System.IObserver<T>! observer) -> void
+ReactiveUI.Primitives.Reactive.Advanced.TaskChainSignal<T>
+ReactiveUI.Primitives.Reactive.Advanced.TaskChainSignal<T>.Subscribe(System.IObserver<T>! observer) -> System.IDisposable!
+ReactiveUI.Primitives.Reactive.Advanced.TaskChainSignal<T>.TaskChainSignal(System.IObservable<System.Threading.Tasks.Task<T>!>! sources) -> void
